added in basic tests for api gateway stack
