+++ test/aws-resource-types.ts
@@ -1,1 +1,9 @@
+// DynamoDB Types
 export const DYNAMODB_TABLE_TYPE = 'AWS::DynamoDB::Table';
+
+// Lambda Types
+export const LAMBDA_FUNCTION_TYPE = 'AWS::Lambda::Function';
+
+// API Gateway Types
+export const REST_API_TYPE = 'AWS::ApiGateway::RestApi';
+export const API_RESOURCE_TYPE = 'AWS::ApiGateway::Resource';
